app/requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
idna==2.8
Markdown==3.1.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.6
xmltodict==0.12.0
